chore: move hardcoded persona details to metadata
